task: test spawn_local tracing task location
This is a new test to make sure that a task created by the task Builder
`spawn_local` function is assigned the correct task location in the
created span.

Right now, this new test is using the `tracing-mock` crate from the
`tracing` repository via a local clone. For this to be merged, it would
be necessary to publish `tracing-mock` to crates.io.